refactor(workflows): isolate PostgreSQL persistence
Keep durable workflow SQL, row mapping, and transaction recovery behind one repository while preserving the existing Database contract.
